[refactor] Make abstract code more generally applicable than just XQuery
