+++ extensions/modules/src/org/exist/xquery/modules/mail/MailSessionFunctions.java
@@ -30,12 +30,12 @@ import javax.mail.Session;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.exist.dom.QName;
+import org.exist.util.ParametersExtractor;
 import org.exist.xquery.BasicFunction;
 import org.exist.xquery.Cardinality;
 import org.exist.xquery.FunctionSignature;
 import org.exist.xquery.XPathException;
 import org.exist.xquery.XQueryContext;
-import org.exist.xquery.modules.ModuleUtils;
 import org.exist.xquery.value.FunctionParameterSequenceType;
 import org.exist.xquery.value.FunctionReturnSequenceType;
 import org.exist.xquery.value.IntegerValue;
@@ -98,7 +98,7 @@ public class MailSessionFunctions extends BasicFunction
 		
 		if( args.length == 1 ) {
 			// try and get the session properties
-			props = ModuleUtils.parseProperties( ((NodeValue) args[0].itemAt(0)).getNode() );
+			props = ParametersExtractor.parseProperties( ((NodeValue) args[0].itemAt(0)).getNode() );
 		}
 		
 		Session session = Session.getInstance( props, null );
