+++ src/org/exist/util/Configuration.java
@@ -24,7 +24,6 @@ import org.apache.logging.log4j.Logger;
 
 import org.exist.repo.Deployment;
 
-import org.exist.xquery.modules.ModuleUtils;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
@@ -547,7 +546,7 @@ public class Configuration implements ErrorHandler
 
             final NodeList nlProperties = htmlToXml.getElementsByTagName(HtmlToXmlParser.HTML_TO_XML_PARSER_PROPERTIES_ELEMENT);
             if(nlProperties.getLength() > 0) {
-                final Properties pProperties = ModuleUtils.parseProperties(nlProperties.item(0));
+                final Properties pProperties = ParametersExtractor.parseProperties(nlProperties.item(0));
                 if(pProperties != null) {
                     final Map<String, Object> properties = new HashMap<>();
                     pProperties.forEach((k,v) -> properties.put(k.toString(), v));
@@ -557,7 +556,7 @@ public class Configuration implements ErrorHandler
 
             final NodeList nlFeatures = htmlToXml.getElementsByTagName(HtmlToXmlParser.HTML_TO_XML_PARSER_FEATURES_ELEMENT);
             if(nlFeatures.getLength() > 0) {
-                final Properties pFeatures = ModuleUtils.parseFeatures(nlFeatures.item(0));
+                final Properties pFeatures = ParametersExtractor.parseFeatures(nlFeatures.item(0));
                 if(pFeatures != null) {
                     final Map<String, Boolean> features = new HashMap<>();
                     pFeatures.forEach((k,v) -> features.put(k.toString(), Boolean.valueOf(v.toString())));
